rename packageUpload to package
